+++ phpBB2/admin/admin_smilies.php
@@ -22,7 +22,6 @@
 /**************************************************************************
 *	This file will be used for modifying the smiley settings for a board.
 **************************************************************************/
-// CTracker_Ignore: File Checked By Human
 // Tell the Security Scanner that reachable code in this file is not a security issue