+++ phpBB2/admin/admin_styles.php
@@ -10,7 +10,6 @@
  *
  *
  ***************************************************************************/
-// CTracker_Ignore: File Checked By Human
 // Tell the Security Scanner that reachable code in this file is not a security issue