+++ phpBB2/includes/smtp.php
@@ -19,7 +19,6 @@
  *
  ***************************************************************************/
 
-// CTracker_Ignore: File Checked By Human
 // Tell the Security Scanner that this constant is allowed
 define('SMTP_INCLUDED', 1);